faith_backend/requirements.txt
django==4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
django-filter
dj-database-url
django-storages
boto3
django-anymail[brevo]
Pillow
gunicorn
requests
django-guardian
